Class design change:To reduce multiple Overide in derived class
