+++ skia/BUILD.gn
@@ -37,11 +37,11 @@ rn_cxx_component("Rnsvg") {
     "react/renderer/components/rnsvg/RNSVGProps.cpp",
     "RSkSVGNode.cpp",
     "RSkSVGNode.h",
-    "RSkSVGContainer.cpp",
-    "RSkSVGContainer.h",
-    "RSkSVGHiddenContainer.h",
-    "RSkSVGShape.cpp",
-    "RSkSVGShape.h",
+    "RSkSVGContainerComponent.cpp",
+    "RSkSVGContainerComponent.h",
+    "RSkSVGHiddenContainerComponent.h",
+    "RSkSVGShapeComponent.cpp",
+    "RSkSVGShapeComponent.h",
     "RSkComponentProviderRNSVGSvgView.cpp",
     "RSkComponentProviderRNSVGSvgView.h",
     "RSkComponentRNSVGSvgView.cpp",
@@ -68,7 +68,6 @@ rn_cxx_component("Rnsvg") {
     "RSkComponentRNSVGRect.h",
     "RSkComponentProviderRNSVGDefs.cpp",
     "RSkComponentProviderRNSVGDefs.h",
-    "RSkComponentRNSVGDefs.cpp",
     "RSkComponentRNSVGDefs.h",
     "RSkComponentProviderRNSVGUse.cpp",
     "RSkComponentProviderRNSVGUse.h",
