+++ skia/RSkComponentRNSVGPath.h
@@ -7,32 +7,25 @@
 
 #pragma once
 
-#include "ReactSkia/components/RSkComponent.h"
-
-#include "RSkSVGShape.h"
+#include "RSkSVGShapeComponent.h"
 
 namespace facebook {
 namespace react {
 
 using namespace std;
 
-class RSkComponentRNSVGPath final : public RSkComponent,public RSkSVGShape{
+class RSkComponentRNSVGPath final : public RSkSVGShapeComponent{
   public:
 
     RSkComponentRNSVGPath(const ShadowView &shadowView);
     ~RSkComponentRNSVGPath() = default;
 
-// Overrides for Base class : RSkComponent
+  // Override for Base class : RSkComponent
   RnsShell::LayerInvalidateMask updateComponentProps(SharedProps newViewProps,bool forceUpdate) override;
-  void mountChildComponent(std::shared_ptr<RSkComponent> newChildComponent,const int index)override;
-  void unmountChildComponent(std::shared_ptr<RSkComponent> oldChildComponent,const int index)override;
 
   protected:
 
-// Overrides for Base class : RSkComponent
-    void OnPaint(SkCanvas *canvas) override {};
-
-// Overrides for Base class : RSkSVGShape
+    // Overrides for Base class : RSkSVGShape
     void onSetAttribute(SkSVGAttribute, const SkSVGValue&) override;
     void onDraw(SkCanvas*, const SkSVGLengthContext&, const SkPaint&,SkPathFillType) const override;
 
@@ -40,7 +33,7 @@ class RSkComponentRNSVGPath final : public RSkComponent,public RSkSVGShape{
 
     mutable SkPath path_; // mutated in onDraw(), to apply inherited fill types.
 
-    typedef RSkSVGShape INHERITED;
+    typedef RSkSVGShapeComponent INHERITED;
 };
 
 } // namespace react
