+++ skia/RSkSVGShapeComponent.cpp
@@ -5,38 +5,43 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include "experimental/svg/model/SkSVGRenderContext.h"
 #include "include/core/SkCanvas.h"
 
-#include "RSkSVGShape.h"
+#include "RSkSVGShapeComponent.h"
 
 using namespace std;
 
 namespace facebook {
 namespace react {
 
-RSkSVGShape::RSkSVGShape(SkSVGTag t) : INHERITED(t) {}
+RSkSVGShapeComponent::RSkSVGShapeComponent(const ShadowView &shadowView,
+                                           RnsShell::LayerType layerType,
+                                           SkSVGTag t)
+    : RSkComponent(shadowView,layerType),
+      RSkSVGNode(t) {}
 
-void RSkSVGShape::onRender(const SkSVGRenderContext& ctx) const {
+void RSkSVGShapeComponent::onRender(const SkSVGRenderContext& ctx) const {
 
   const auto fillType = ctx.presentationContext().fInherited.fFillRule->asFillType();
 
-  if (const SkPaint* fillPaint = ctx.fillPaint()) {
+  if (SkPaint* fillPaint = const_cast<SkPaint*>(ctx.fillPaint())) {
     SkColor color=fillPaint->getColor();
     RNS_LOG_DEBUG("FillColor : "<<color);
     RNS_LOG_DEBUG("A : "<<SkColorGetA(color));
     RNS_LOG_DEBUG("R : "<<SkColorGetR(color));
     RNS_LOG_DEBUG("G : "<<SkColorGetG(color));
     RNS_LOG_DEBUG("B : "<<SkColorGetB(color));
+    fillPaint->setAntiAlias(true);
     this->onDraw(ctx.canvas(), ctx.lengthContext(), *fillPaint, fillType);
   }
-  if (const SkPaint* strokePaint = ctx.strokePaint()) {
+  if (SkPaint* strokePaint = const_cast<SkPaint*>(ctx.strokePaint())) {
     SkColor color=strokePaint->getColor();
     RNS_LOG_DEBUG("StrokeColor : "<<color);
     RNS_LOG_DEBUG("A : "<<SkColorGetA(color));
     RNS_LOG_DEBUG("R : "<<SkColorGetR(color));
     RNS_LOG_DEBUG("G : "<<SkColorGetG(color));
     RNS_LOG_DEBUG("B : "<<SkColorGetB(color));
+    strokePaint->setAntiAlias(true);
     this->onDraw(ctx.canvas(), ctx.lengthContext(), *strokePaint, fillType);
   }
 
@@ -47,6 +52,19 @@ void RSkSVGShape::onRender(const SkSVGRenderContext& ctx) const {
   RNS_LOG_INFO(" getTranslateY "<<matrix.getTranslateY());
   RNS_LOG_INFO(" getSkewX "<<matrix.getSkewX());
   RNS_LOG_INFO(" getSkewY "<<matrix.getSkewY());
+}
+
+
+void RSkSVGShapeComponent::mountChildComponent(std::shared_ptr<RSkComponent> newChildComponent,const int index) {
+  RNS_LOG_INFO("SVG Element of type Shapes can't have child Node :: "<<newChildComponent->getComponentData().componentName);
+}
+
+void RSkSVGShapeComponent::unmountChildComponent(std::shared_ptr<RSkComponent> oldChildComponent,const int index) {
+  RNS_LOG_INFO(" SVG Element Of type Shapes can't have child :: "<<oldChildComponent->getComponentData().componentName);
+}
+
+void RSkSVGShapeComponent::OnPaint(SkCanvas *canvas) {
+  RNS_LOG_INFO(" Paint for Shapes Elements handled via RNSSVG element ");
 }
 
 } // namespace react
